Use pure reflection (cached) instead of MMR for event polymorphism
