--- src/NServiceBus.SqlServer/PubSub/Topics/BasicTopicManager.cs
@@ -1,13 +1,0 @@
-﻿namespace NServiceBus.Transport.SQLServer
-{
-    using System;
-    using System.Collections.Generic;
-
-    class BasicTopicManager : ITopicManager
-    {
-        public IEnumerable<string> GetTopicsFor(Type messageType)
-        {
-            yield return TopicName.From(messageType);
-        }
-    }
-}
