--- src/NServiceBus.SqlServer/PubSub/Topics/ITopicManager.cs
@@ -1,10 +1,0 @@
-﻿namespace NServiceBus.Transport.SQLServer
-{
-    using System;
-    using System.Collections.Generic;
-
-    interface ITopicManager
-    {
-        IEnumerable<string> GetTopicsFor(Type messageType);
-    }
-}
